requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
galileo[langchain]>=1.37.0
openai>=1.0.0
python-dotenv>=1.0.0
tavily-python>=0.4.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0

# RAG Infrastructure
pinecone-client>=3.0.0
langchain-pinecone>=0.0.1
langchain-community>=0.0.10

# Document Processing
pypdf2>=3.0.0
python-docx>=1.1.0
unstructured>=0.11.0
python-magic>=0.4.27

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Authentication
python-multipart>=0.0.6
aiosqlite>=0.19.0 
